+++ scripts/restore.sh
@@ -192,7 +192,7 @@ restore_active_and_alternate_sessions() {
 
 main() {
 	if supported_tmux_version_ok && check_saved_session_exists; then
-		start_spinner
+		start_spinner "Restoring..." "Tmux restore complete!"
 		restore_all_panes
 		restore_pane_layout_for_each_window >/dev/null 2>&1
 		restore_all_pane_processes