+++ scripts/spinner_helpers.sh
@@ -1,5 +1,5 @@
 start_spinner() {
-	$CURRENT_DIR/tmux_spinner.sh "Restoring tmux..." "Tmux restore complete!" &
+	$CURRENT_DIR/tmux_spinner.sh "$1" "$2" &
 	export SPINNER_PID=$!
 }