+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ### master
+- show spinner during the save process
 
 ### v0.4.0, 2014-08-29
 - change plugin name to `tmux-resurrect`. Change all the variable names.